+++ ChangeLog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 
+### Fixed
+- Correct blend mode for transmission in URP
+
 ## [4.4.6] - 2021-11-10
 ### Added
 - Added warning when more than two UV sets are supposed to be imported (not supported yet)